LPS-60935 move
