+++ modules/portal/portal-rtl-servlet/src/main/java/com/liferay/portal/rtl/servlet/RTLServletTracker.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.language.servlet.filter;
+package com.liferay.portal.rtl.servlet;
 
 import java.util.Hashtable;
